mubench.pipeline/requirements.txt
PyYaml
requests
psutil
hurry.filesize
